godep: Update cloudfoundry/noaa
So that we have a newer version of the events package which exposes
`Envelope.Get{Job,Index,Ip}()`. We want these so that we can identify which
VM a metric is coming from.

This version of `cloudfoundry/noaa` is the last before events was split to
`cloudfoundry/sonde-go` and many other things were refactored, including the
removal of heartbeat metrics. I'm not going with the latest version because
I don't want to change too many things at once.