Add prefixJob option
Which prepend the CloudFoundry job name (e.g. `runner_z1`) and job index
(e.g. `0`) to every metric name. We want this in order to debug issues with
particular servers, for example if Doppler is sending messages on one side
but they aren't all being received on the other end.

The way that this option is passed through is sub-optimal. We did consider
some other options like whether we could change the prefix of the statsd
client on the fly, but it wasn't possible/efficient. The codebase could
benefit from some refactoring, to reduce duplication.

To use this you will need to run something like:

    cf set-env graphite-nozzle PREFIX_JOB true