Migrate vendored dependencies to vendor/
Go 1.6+ looks for dependencies here by default without needing to modify
the GOPATH. godep will be deprecating the `Godeps/_workspace` directory
once Go 1.8 is released.

This resaves all the vendored dependencies to vendor/ which is the new
default with Go 1.6. This has been done following the process described
here: https://github.com/tools/godep#migrating-to-vendor

Note: this has been done using the current version of godep, which is
more efficient about what it selects to vendor. Specifically, it
excludes all test files, and testdata directories. This is why there are
lots of deleted files in this commit.